remove the Error type alias.
I've received feedback that it's confusing!
Two reasons: it's just awful to have to talk about "error" vs "Error"
(differing only by capitalization); and also, because something in the
waters of golang conventions makes people suspect that name should
contain a constructor function(!).

It also simply isn't expected to be often used, so it's very
undeserving of taking such a prominent piece of namespace.

So: begone.