the untyped nil problem made me very angry today.
I've implemented defense-in-depth checks in several places.

It bothers me that this causes us to have a dependency on the reflect package...
but this problem is simply too ridiculous and too infuriating to allow in the wild.
Getting a nil pointer dereference that causes badly broken output in the middle
of formating an error message or producing the json to describe an error
is simply unacceptable redoubling of problems under already-unhappy circumstances,
and it's worth a great deal of effort to avoid it at all costs.