Change the behavior of serum.Message to fallback to Error string on non-serum types.
This seems to produce slightly less surprising behavior in the wild;
and it means less information is discarded, which in any scenario that
might contain a flicker of doubt, is generally the right way to go.

This also means that when Standardize functions are used (or, the
Errorf helper constructor, which uses Standardize in turn) on non-serum
error values, they keep their messages better.  Big improvement.